feat(DEVEXP-36): split Numbers API related files into smaller units
