+++ sinch/domains/numbers/endpoints/active/get_number_configuration.py
@@ -1,17 +1,17 @@
 from sinch.core.models.http_response import HTTPResponse
 from sinch.domains.numbers.endpoints.numbers_endpoint import NumbersEndpoint
 from sinch.core.enums import HTTPAuthentication, HTTPMethods
-from sinch.domains.numbers.models.responses import UpdateNumberConfigurationResponse
-from sinch.domains.numbers.models.requests import UpdateNumberConfigurationRequest
 
+from sinch.domains.numbers.models.active.requests import GetNumberConfigurationRequest
+from sinch.domains.numbers.models.active.responses import GetNumberConfigurationResponse
 
-class UpdateNumberConfigurationEndpoint(NumbersEndpoint):
+class GetNumberConfigurationEndpoint(NumbersEndpoint):
     ENDPOINT_URL = "{origin}/v1/projects/{project_id}/activeNumbers/{phone_number}"
-    HTTP_METHOD = HTTPMethods.PATCH.value
+    HTTP_METHOD = HTTPMethods.GET.value
     HTTP_AUTHENTICATION = HTTPAuthentication.OAUTH.value
 
-    def __init__(self, project_id: str, request_data: UpdateNumberConfigurationRequest):
-        super(UpdateNumberConfigurationEndpoint, self).__init__(project_id, request_data)
+    def __init__(self, project_id: str, request_data: GetNumberConfigurationRequest):
+        super(GetNumberConfigurationEndpoint, self).__init__(project_id, request_data)
         self.project_id = project_id
         self.request_data = request_data
 
@@ -22,13 +22,9 @@ class UpdateNumberConfigurationEndpoint(NumbersEndpoint):
             phone_number=self.request_data.phone_number
         )
 
-    def request_body(self):
-        self.request_data.phone_number = None
-        return self.request_data.as_json()
-
-    def handle_response(self, response: HTTPResponse) -> UpdateNumberConfigurationResponse:
-        super(UpdateNumberConfigurationEndpoint, self).handle_response(response)
-        return UpdateNumberConfigurationResponse(
+    def handle_response(self, response: HTTPResponse) -> GetNumberConfigurationResponse:
+        super(GetNumberConfigurationEndpoint, self).handle_response(response)
+        return GetNumberConfigurationResponse(
             phone_number=response.body["phoneNumber"],
             project_id=response.body["projectId"],
             display_name=response.body["displayName"],
