+++ sinch/domains/numbers/endpoints/available/activate_number.py
@@ -1,8 +1,8 @@
 from sinch.core.models.http_response import HTTPResponse
 from sinch.domains.numbers.endpoints.numbers_endpoint import NumbersEndpoint
 from sinch.core.enums import HTTPAuthentication, HTTPMethods
-from sinch.domains.numbers.models.responses import ActivateNumberResponse
-from sinch.domains.numbers.models.requests import ActivateNumberRequest
+from sinch.domains.numbers.models.available.requests import ActivateNumberRequest
+from sinch.domains.numbers.models.available.responses import ActivateNumberResponse
 
 
 class ActivateNumberEndpoint(NumbersEndpoint):
