+++ sinch/domains/numbers/endpoints/available/search_for_number.py
@@ -1,9 +1,8 @@
 from sinch.core.models.http_response import HTTPResponse
 from sinch.domains.numbers.endpoints.numbers_endpoint import NumbersEndpoint
 from sinch.core.enums import HTTPAuthentication, HTTPMethods
-from sinch.domains.numbers.models.responses import CheckNumberAvailabilityResponse
-from sinch.domains.numbers.models.requests import CheckNumberAvailabilityRequest
-
+from sinch.domains.numbers.models.available.responses import CheckNumberAvailabilityResponse
+from sinch.domains.numbers.models.available.requests import CheckNumberAvailabilityRequest
 
 class SearchForNumberEndpoint(NumbersEndpoint):
     ENDPOINT_URL = "{origin}/v1/projects/{project_id}/availableNumbers/{phone_number}"
