+++ sinch/domains/numbers/models/regions/requests.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+from sinch.core.models.base_model import SinchRequestBaseModel
+
+
+@dataclass
+class ListAvailableRegionsForProjectRequest(SinchRequestBaseModel):
+    number_type: str
+    number_types: list
